feat: stage SIEM MVP pilot — agents, hunt approvals, soak ops, handoff
Land the staging/siem-mvp worktree: enrollment/credential rotation, HNT-007
approval path, SIEM-009 backup/soak/Object Lock tooling, frontend-v3 +
backend/agent pipeline alignment, and AI handoff docs (including full-stack
audit prompt). Secrets stay gitignored (ADMIN_BOOTSTRAP, .env).

+++ backend/src/main/java/com/hivearmor/aop/pointcut/AlertPointcut.java
@@ -16,7 +16,7 @@ public class AlertPointcut {
         // Method is empty as this is just a Pointcut, the implementations are in the advices.
     }
 
-    public void convertToIncidentPointcut(Query query, String incidentName, Integer incidentId, Instant incidentCreationDate, String incidentCreatedBy, String incidentSource, String indexPattern) {
+    public void convertToIncidentPointcut(Query query, String incidentName, long incidentId, Instant incidentCreationDate, String incidentCreatedBy, String incidentSource, String indexPattern) {
         // Method is empty as this is just a Pointcut, the implementations are in the advices.
     }
 }
